SEPP-34 Move user login and registration functionality to /api/user
