+++ dating_app/main.py
@@ -1,25 +1,17 @@
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-from fastapi import APIRouter, FastAPI, Form, Request, Response, status
+from fastapi import APIRouter, FastAPI, Form, Request, status
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 from pydantic import EmailStr
 
 import dating_app.services as services
 from dating_app.api.chat import chat
+from dating_app.api.user import user
 from dating_app.core import websocket
 from dating_app.db import mongo
-from dating_app.schemas.User import (
-    LoginResponse,
-    LoginResponseBase,
-    RegisterResponse,
-    RegisterResponseBase,
-    UserCreate,
-    UserModel,
-    UserPublic,
-    UserSearch,
-)
+from dating_app.schemas.User import UserPublic, UserSearch
 
 BASE_PATH = Path(__file__).resolve().parent
 templates = Jinja2Templates(directory=BASE_PATH / "templates")
@@ -57,60 +49,6 @@ async def read_item(request: Request, item_id: int):
 @api_router.get("/searchpage", status_code=200)
 async def search_page(request: Request):
     return templates.TemplateResponse("search.html", {"request": request})
-
-
-@api_router.post(
-    "/login",
-    status_code=status.HTTP_200_OK,
-    response_model=LoginResponse,
-    responses=LoginResponseBase.Config.schema_extra,  # type: ignore
-)
-async def login(
-    response: Response,
-    email: str = Form(...),
-    password: str = Form(...),
-):
-
-    user_form = UserModel(email=email, password=password)
-    login_response = await services.authenticate_user(mongo.mongodb, user_form)
-    print(f"user is {login_response}")
-    if "Invalid" in login_response.message:
-        print(f"Invalid credentials 401: {email}")
-        response.status_code = status.HTTP_401_UNAUTHORIZED
-    else:
-        response.set_cookie(key="X-Authorization", value=email, httponly=True)
-
-    return login_response
-
-
-@api_router.post("/logout", status_code=status.HTTP_200_OK)
-async def logout(response: Response):
-    logout_response = await services.logout(mongo.mongodb)
-    print(logout_response)
-    return logout_response
-
-
-@api_router.post(
-    "/register",
-    status_code=status.HTTP_201_CREATED,
-    response_model=RegisterResponse,
-    responses=RegisterResponseBase.Config.schema_extra,  # type: ignore
-)
-async def register(
-    response: Response,
-    email: str = Form(...),
-    password: str = Form(...),
-    confirmPassword: str = Form(...),
-):
-    user = UserCreate(
-        email=email, password=password, confirmed_password=confirmPassword
-    )
-    register_response = await services.create_user(mongo.mongodb, user)
-
-    if "already in use." in register_response.message:
-        response.status_code = status.HTTP_409_CONFLICT
-
-    return register_response
 
 
 @api_router.get(
@@ -204,6 +142,7 @@ async def update_user_me(
 
 
 api_router.include_router(websocket.router)
+api_router.include_router(user.router, prefix="/api/user", tags=["user"])
 api_router.include_router(chat.router, prefix="/api/chat", tags=["chat"])
 app.include_router(api_router)
